reduced code duplication
